a) Enhancement: tons of changes to genericize the reading of JSON responses and transforming them to type-specific beans.
b) Bug Fix: when AuthCache.getApiKey() was invoked with an empty cache, it was producing the wrong API Key. Root cause was that getHashedPassword() was getting called before getToken(). However the hashed password doesn't get stoted until getToken() is successful, so getHashedPassword() was returning null on the very first call. This resulted in the incorrect API key.